feat: 글 삭제 confirm 추가

+++ src/components/freeboard/FreeBoardDeleteView.vue
@@ -13,11 +13,14 @@ export default {
         };
     },
     created() {
-        console.log(this.$route.params.boardno);
-        http.delete(`/freeboard/${this.$route.params.boardno}`).then(({ data }) => {
-            this.board = data
-        })
-        this.$router.push({ name: "freeboardList" });
+        if (confirm('정말 글을 삭제하시겠습니까?')) {
+            console.log(this.$route.params.boardno);
+            http.delete(`/freeboard/${this.$route.params.boardno}`).then(({ data }) => {
+                this.board = data
+            })
+            alert("글이 삭제되었습니다.")
+            this.$router.push({ name: "freeboardList" });
+        } else this.$router.back()
     },
     methods: {},
 };
